Class_JS/class07/drf_backend/drf_backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["asd <asd@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.5"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
